+++ src/procedure.ts
@@ -24,15 +24,8 @@ export abstract class Procedure extends BirbsRunnable {
    * @abstract
    * @warning Do not use this method yourself! It may introduce a state inconsistency to
    * your context.
+   *
+   * Remember to define the type of the descriptable if you'e useing typescript
    */
-  public abstract execute (context : Context) : Promise<void>;
-
-  /**
-   * It is the method used when the procedure is to be executed
-   * @async
-   * @abstract
-   * @warning Do not use this method yourself! It may introduce a state inconsistency to
-   * your context.
-   */
-  public abstract execute (context : Context, descriptable?) : Promise<void>;
+  public abstract execute (context : Context, descriptable ?: unknown) : Promise<void>;
 }